fix: Adds failed job count

+++ pkg/inventory/api/resource_type.go
@@ -43,8 +43,8 @@ type ListResourceTypeCompositionResponse struct {
 
 type ResourceTypeTrendDatapoint struct {
 	Count                                   int       `json:"count" example:"100" minimum:"0"`
-	TotalDescribedConnectionCount           int       `json:"totalConnectionCount"`
-	TotalSuccessfulDescribedConnectionCount int       `json:"totalSuccessfulDescribedConnectionCount"`
+	TotalDescribedConnectionCount           int64     `json:"totalConnectionCount"`
+	TotalSuccessfulDescribedConnectionCount int64     `json:"totalSuccessfulDescribedConnectionCount"`
 	Date                                    time.Time `json:"date" format:"date-time"`
 }
